converted totalspend amount from int to double in TransactionAnalysisResult model

+++ bettrsdkapi/src/main/java/com/mvalu/bettr_api/transactions/TransactionAnalysisResult.kt
@@ -8,7 +8,7 @@ class TransactionAnalysisResult {
     var totalAmountSpend: Double? = null
 
     @field:Json(name = "totalSpend")
-    var totalSpend: Int? = null
+    var totalSpend: Double? = null
 
     @field:Json(name = "highestSpendValue")
     var highestSpendValue: Double? = null
